fix: 플레이어 배팅 결과 기능 수정

+++ src/main/java/domain/blackJack/BlackJack.java
@@ -37,7 +37,7 @@ public class BlackJack {
         dealer.draw(standard);
     }
 
-    public Map<Player, Double> calculatePlayerResult() {
+    public Map<Player, Integer> calculatePlayerResult() {
         return players.calculateWinner(dealer);
     }
 }
